Brackets!
I have added the ==num in row two as the logic will need to be either 1 (commissioned care) or null.

The reason the code will not clear the DP_Reviewneeded field is due to the placement of the brackets. I have moved a set of brackets from line one and moved the starting bracket to line 3 and the closing bracket to line 5.